requirements.txt
 django
 djangorestframework
 django-admin-interface
 django-cors-headers
 mysqlclient